6_use_cases/usecases/llm_customization_job_governance/smtj_batch_utils/requirements.txt
boto3>=1.42.76
botocore>=1.42.76
pyyaml
sagemaker>=3.7.0